backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
black==23.7.0
botocore==1.31.28
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
deprecation==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
frozenlist==1.4.0
gotrue==1.0.3
gql==3.4.1
graphql-core==3.2.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
jmespath==1.0.1
Markdown==3.4.4
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.8
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
postgrest==0.10.8
psycopg==3.1.10
psycopg-binary==3.1.10
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
realtime==1.0.0
requests==2.31.0
requests-toolbelt==0.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
websockets==10.4
yarl==1.9.2
